Make mygame_teardown only go up one level
The specs for Command (in spec/lib/entityjs/command_spec.rb) were
going up two levels with mygame_teardown, because it's run twice in a
row with before(:all).

This would either: fail, when the parent directory contained a scripts
directory; or write to ../build/ when it didn't. Neither is desirable.